pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embodied-datakit"
version = "0.1.0"
description = "Dataset compiler for heterogeneous robot trajectories (RLDS, LeRobot, Open X-Embodiment)"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "EmbodiedDataKit Contributors" }
]
keywords = [
    "robotics",
    "datasets",
    "rlds",
    "lerobot",
    "open-x-embodiment",
    "machine-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
tfds = [
    "tensorflow-datasets>=4.9",
    "tensorflow>=2.15",
]
video = [
    "ffmpeg-python>=0.2",
    "av>=11.0",
]
query = [
    "duckdb>=0.9",
    "polars>=0.20",
]
torch = [
    "torch>=2.0",
]
rlbench = [
    "pyrep>=4.1",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "hypothesis>=6.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]
all = [
    "embodied-datakit[tfds,video,query,torch,dev,docs]",
]

[project.scripts]
edk = "embodied_datakit.cli:main"

[project.urls]
Homepage = "https://github.com/embodied-datakit/embodied-datakit"
Documentation = "https://embodied-datakit.github.io/"
Repository = "https://github.com/embodied-datakit/embodied-datakit"
Issues = "https://github.com/embodied-datakit/embodied-datakit/issues"

[tool.hatch.build.targets.wheel]
packages = ["embodied_datakit", "cli"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict=
]

[tool.ruff.isort]
known-first-party = ["embodied_datakit"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tensorflow.*",
    "tensorflow_datasets.*",
    "av.*",
    "ffmpeg.*",
    "pyrep.*",
    "rlbench.*",
    "duckdb.*",
    "polars.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_tfds: marks tests requiring TensorFlow Datasets",
    "requires_video: marks tests requiring video dependencies",
]

[tool.coverage.run]
source = ["embodied_datakit"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
